Split out metadata_json module
