+++ src/providers/mod.rs
@@ -11,6 +11,7 @@ pub mod haskell;
 pub mod launchpad;
 pub mod maven;
 pub mod meson;
+pub mod metadata_json;
 pub mod metainfo;
 pub mod nuspec;
 #[cfg(feature = "opam")]
